refactor(rust-api): extract auth/access middleware into auth.rs (sc-1636)
First production-seam split of the API lib.rs. Move the access-token
middleware cluster (access_control, cors_layer, requires_token,
is_authorized, token_from_headers, constant_time_eq) into a dedicated
auth.rs child module.

Pattern: auth.rs starts with `use super::*;` so it sees the crate
root's items (AppState, Settings, PUBLIC_PATHS, axum/tower imports)
unchanged; the three fns referenced from outside the module
(access_control + cors_layer by create_app, is_authorized by
verify_access) are pub(crate) and re-imported at the root, so all
call sites stay unqualified. tests.rs's requires_token import is
repointed to super::auth. Behavior-preserving.
