+++ apps/rust-api/src/lib.rs
@@ -62,6 +62,9 @@ use tokio_util::io::ReaderStream;
 use tower_http::cors::{AllowOrigin, CorsLayer};
 use uuid::Uuid;
 
+mod auth;
+use auth::{access_control, cors_layer, is_authorized};
+
 const PUBLIC_PATHS: &[&str] = &[
     "/api/v1/health",
     "/api/v1/access",
@@ -5107,99 +5110,6 @@ fn sse_message_event(message: EventMessage) -> Event {
 
 fn heartbeat_event() -> Event {
     Event::default().event("heartbeat").data(HEARTBEAT_SSE_DATA)
-}
-
-async fn access_control(
-    State(state): State<AppState>,
-    request: Request<axum::body::Body>,
-    next: Next,
-) -> Response {
-    if request.method() == Method::OPTIONS
-        || !requires_token(request.uri().path())
-        || is_authorized(request.headers(), &state.settings)
-    {
-        return next.run(request).await;
-    }
-
-    (
-        StatusCode::UNAUTHORIZED,
-        Json(json!({
-            "detail": "SceneWorks access token required",
-            "authRequired": true
-        })),
-    )
-        .into_response()
-}
-
-fn cors_layer(settings: &Settings) -> CorsLayer {
-    let origins = settings
-        .cors_origins
-        .iter()
-        .filter_map(|origin| HeaderValue::from_str(origin).ok())
-        .collect::<Vec<_>>();
-
-    CorsLayer::new()
-        .allow_origin(AllowOrigin::list(origins))
-        .allow_methods([
-            Method::GET,
-            Method::POST,
-            Method::PUT,
-            Method::PATCH,
-            Method::DELETE,
-            Method::OPTIONS,
-        ])
-        .allow_headers([
-            header::AUTHORIZATION,
-            header::CONTENT_TYPE,
-            HeaderName::from_static("x-sceneworks-token"),
-        ])
-}
-
-/// Whether a path is gated by the access token. Only `/api/*` routes are
-/// protected (minus the explicitly public ones); everything else is the
-/// embedded web bundle / SPA fallback, which a browser must be able to load
-/// before it can attach the token header.
-fn requires_token(path: &str) -> bool {
-    path.starts_with("/api/") && !PUBLIC_PATHS.contains(&path)
-}
-
-fn is_authorized(headers: &HeaderMap, settings: &Settings) -> bool {
-    if settings.access_token.is_empty() {
-        return true;
-    }
-    constant_time_eq(
-        token_from_headers(headers).as_bytes(),
-        settings.access_token.as_bytes(),
-    )
-}
-
-fn token_from_headers(headers: &HeaderMap) -> String {
-    if let Some(token) = headers
-        .get("x-sceneworks-token")
-        .and_then(|value| value.to_str().ok())
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-    {
-        return token.to_owned();
-    }
-    headers
-        .get("authorization")
-        .and_then(|value| value.to_str().ok())
-        .map(str::trim)
-        .and_then(|value| value.strip_prefix("Bearer "))
-        .map(str::trim)
-        .unwrap_or_default()
-        .to_owned()
-}
-
-fn constant_time_eq(left: &[u8], right: &[u8]) -> bool {
-    if left.len() != right.len() {
-        return false;
-    }
-    left.iter()
-        .zip(right.iter())
-        .fold(0, |difference, (left, right)| difference | (left ^ right))
-        == 0
 }
 
 async fn store_call<T, F>(state: AppState, operation: F) -> Result<T, ApiError>
